requirements.txt
torch==1.11.0
torchvision==0.12.0
scikit-learn~=1.1.3
scikit-image~=0.19.1
numpy~=1.21.5
Pillow~=9.0.1
scipy~=1.9.3
yamldataclassconfig~=1.5.0
EasyDict~=1.10
matplotlib~=3.6.2
xgboost~=1.7.1
PyYAML~=5.4.1
pandas~=1.5.2
hyperopt~=0.2.7
seaborn
mlflow
torchviz
kornia
torchtoolbox